Added rugaps patch.
Patch link:
http://dwm.suckless.org/patches/ru_gaps/

Description:
Runtime useless gaps

This patch adds gaps like tilegap that are resizeable at runtime like fullgaps, but like uselessgap it removes gaps and borders when in monocle mode.

This patch also incorporates the noborder patch, but without the check for 'is not fullscreen', to make it work properly with fakefullscreen.